Use our old CMake file
Upstream now switched to CMake finally,
but it does not compile shared code with a shared library.
And it messed up intermediate and install directories.
Stay unchanged until it addressed.

* Don't link ncurses

ncurses comes with several variants (e.g. libncurses{,w}),
with the *w variant is the superset.
For this reason perhaps, Alpine Linux only have the *w variant.

Anyway we don't use that directly, it merely an indirect
dependence from readline.